pyproject.toml
[project]
name = "social-media-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[google-genai]>=1.9.3",
    "crewai-tools>=1.9.3",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
]
